backend_python/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
torch
torchaudio
numpy
openai-whisper
ffmpeg-python
soundfile
tqdm
requests
python-dotenv

